// src/app/success-message/models/PbResponseDataModel.ts
export class PbResponseDataModel {
  utf8?: string;
  auth_cv_result?: string;
  req_locale?: string;
  req_payer_authentication_indicator?: string;
  payer_authentication_acs_transaction_id?: string;
  auth_trans_ref_no?: string;
  req_card_type_selection_indicator?: string;
  payer_authentication_enroll_veres_enrolled?: string;
  req_bill_to_surname?: string;
  req_card_expiry_date?: string;
  card_type_name?: string;
  auth_amount?: string;
  auth_response?: string;
  bill_trans_ref_no?: string;
  req_payment_method?: string;
  req_payer_authentication_merchant_name?: string;
  auth_time?: string;
  transaction_id?: string;
  req_card_type?: string;
  payer_authentication_transaction_id?: string;
  payer_authentication_pares_status?: string;
  payer_authentication_cavv?: string;
  auth_avs_code?: string;
  auth_code?: string;
  payer_authentication_specification_version?: string;
  req_bill_to_address_country?: string;
  auth_cv_result_raw?: string;
  req_profile_id?: string;
  signed_date_time?: string;
  req_bill_to_address_line1?: string;
  payer_authentication_validate_e_commerce_indicator?: string;
  req_card_number?: string;
  signature?: string;
  req_bill_to_address_city?: string;
  auth_cavv_result?: string;
  reason_code?: string;
  req_bill_to_forename?: string;
  req_payer_authentication_acs_window_size?: string;
  payment_account_reference?: string;
  request_token?: string;
  auth_cavv_result_raw?: string;
  req_amount?: string;
  req_bill_to_email?: string;
  payer_authentication_reason_code?: string;
  auth_avs_code_raw?: string;
  req_currency?: string;
  decision?: string;
  message?: string;
  signed_field_names?: string;
  req_transaction_uuid?: string;
  payer_authentication_eci?: string;
  req_transaction_type?: string;
  payer_authentication_xid?: string;
  req_access_key?: string;
  req_reference_number?: string;
  payer_authentication_validate_result?: string;
  auth_reconciliation_reference_number?: string;
}